requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cloudinary==1.28.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-heroku==0.0.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-environ==0.4.54
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
whitenoise==5.3.0
